feat: create consolidated code generator
Often when accessing GPT with code it is useful to include the relevant
code that is being discussed. The `util generate-prompt` command will
consolidate files in a given directory into a single string that can be
included as context.

The prompt has an ignore file, `.gptignore` which will be used to
determine which files to skip when generating a prompt.